rename typing
Turns out typing collides with a system package. Rename it to not
collide
